api/requirements.txt
django>=4.1
psycopg2>=2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2 
django-cors-headers==4.1.0 
# python-ipware==1.0.3 
django-ip-geolocation==1.6.1 
pillow==10.0.1 
# django-filter==23.2
# drf-spectacular==0.24.2 
# pandas==1.5.2
